srd5.1 data pack implementation WIP, we're back on track!
