+++ src/server/data/presets/dnd5eSrd/README.md
@@ -1,0 +1,63 @@
+# D&D 5E SRD 5.1 Data Pack
+
+> This work includes material taken from the System Reference Document 5.1 (“SRD 5.1”) by Wizards of 
+the Coast LLC and available at https://dnd.wizards.com/resources/systems-reference-document. The 
+SRD 5.1 is licensed under the Creative Commons Attribution 4.0 International License available at 
+https://creativecommons.org/licenses/by/4.0/legalcode.
+
+# Models Used
+
+## Extends CustomBaseDocument
+
+- [x] Game
+- [x] Product
+- [x] Dice
+
+## Extends ContentBaseDocument, smaller amounts of data 
+
+- [ ] ActionType
+- [ ] CastingComponent
+- [ ] SpellSchool
+- [ ] SpellLevel
+- [x] AttackType
+- [ ] DamageType
+- [ ] Condition
+- [ ] Alignment
+- [ ] Ability
+- [ ] Skill
+- [ ] Currency
+- [ ] CurrencyExchange
+- [ ] Language
+- [ ] LevelledProficiencyBonusTable
+- [ ] Sense
+- [ ] Size
+- [ ] Species
+- [ ] SubcreatureCategory
+- [ ] Universe
+- [ ] Place
+
+## Extends ContentBaseDocument, does not use SRD-specific data
+
+- [ ] Day
+- [ ] Month
+- [ ] Epoch
+- [ ] Calendar
+
+## Extends ContentBaseDocument, larger amounts of data 
+
+- [ ] Spell 
+- [ ] Item
+- [ ] CreatureInstance
+- [ ] HeroClass
+- [ ] Feature
+
+# Models Not Used
+
+## Extends ContentBaseDocument, but NOT needed
+
+- [x] Encounter
+- [x] Lore
+
+## Extends CustomBaseDocument, but NOT needed
+
+- [x] Campaign
