+++ src/server/functions/serverUtils.js
@@ -4,7 +4,7 @@
 const {databaseConnector} = require('../database');
 const { User } = require('../models/extendsDocument/UserModel');
 const {ServerConfig} = require('../models/extendsDocument/ServerConfig');
-const { createDefaultData } = require('../data/presets/dnd5e');
+const { createDefaultData } = require('../data/presets/dnd5eSrd');
 const modelUtils = require('./modelUtils');
 let db = null;
 const databaseInitCheck = async () => {
